--- a/.gitmodules
+++ b/.gitmodules
@@ -7,9 +7,6 @@
 [submodule "compiler+runtime/third-party/bpptree"]
 	path = compiler+runtime/third-party/bpptree
 	url = https://github.com/jank-lang/BppTree.git
-[submodule "compiler+runtime/third-party/fmt"]
-	path = compiler+runtime/third-party/fmt
-	url = https://github.com/jank-lang/fmt.git
 [submodule "compiler+runtime/third-party/immer"]
 	path = compiler+runtime/third-party/immer
 	url = https://github.com/jank-lang/immer.git