--- a/compiler+runtime/include/cpp/jank/runtime/convert.hpp
+++ b/compiler+runtime/include/cpp/jank/runtime/convert.hpp
@@ -33,7 +33,3 @@ namespace jank::runtime
     }
   };
 }
-
-#include <jank/runtime/convert/into.hpp>
-#include <jank/runtime/convert/from.hpp>
-#include <jank/runtime/convert/function.hpp>