--- a/compiler+runtime/include/cpp/jank/runtime/convert/function.hpp
+++ b/compiler+runtime/include/cpp/jank/runtime/convert/function.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <jank/runtime/convert.hpp>
+#include <jank/runtime/convert/into.hpp>
 
 namespace jank::runtime
 {