--- a/compiler+runtime/include/cpp/jank/runtime/core/seq_ext.hpp
+++ b/compiler+runtime/include/cpp/jank/runtime/core/seq_ext.hpp
@@ -3,6 +3,7 @@
 #include <jank/runtime/object.hpp>
 #include <jank/runtime/visit.hpp>
 #include <jank/runtime/behavior/seqable.hpp>
+#include <jank/runtime/core/equal.hpp>
 
 namespace jank::runtime
 {