--- a/compiler+runtime/include/cpp/jank/util/string_builder.hpp
+++ b/compiler+runtime/include/cpp/jank/util/string_builder.hpp
@@ -21,13 +21,15 @@ namespace jank::util
     string_builder &operator=(string_builder &&) = delete;
 
     string_builder &operator()(native_bool d) &;
-    string_builder &operator()(native_integer d) &;
-    string_builder &operator()(native_real d) &;
+    string_builder &operator()(float d) &;
+    string_builder &operator()(double d) &;
     string_builder &operator()(native_hash d) &;
     string_builder &operator()(void const *d) &;
     string_builder &operator()(int d) &;
     string_builder &operator()(long d) &;
-    string_builder &operator()(size_t d) &;
+    string_builder &operator()(long long d) &;
+    string_builder &operator()(unsigned long d) &;
+    string_builder &operator()(unsigned long long d) &;
     string_builder &operator()(char d) &;
     string_builder &operator()(char32_t d) &;
     string_builder &operator()(char const *d) &;
@@ -35,13 +37,15 @@ namespace jank::util
     string_builder &operator()(native_persistent_string const &d) &;
 
     void push_back(native_bool d) &;
-    void push_back(native_integer d) &;
-    void push_back(native_real d) &;
+    void push_back(float d) &;
+    void push_back(double d) &;
     void push_back(native_hash d) &;
     void push_back(void const *d) &;
     void push_back(int d) &;
     void push_back(long d) &;
-    void push_back(size_t d) &;
+    void push_back(long long d) &;
+    void push_back(unsigned long d) &;
+    void push_back(unsigned long long d) &;
     void push_back(char d) &;
     void push_back(char32_t d) &;
     void push_back(char const *d) &;