--- a/compiler+runtime/src/cpp/jank/compiler_native.cpp
+++ b/compiler+runtime/src/cpp/jank/compiler_native.cpp
@@ -1,10 +1,12 @@
 #include <jank/compiler_native.hpp>
-#include <jank/runtime/convert.hpp>
+#include <jank/runtime/convert/function.hpp>
 #include <jank/runtime/context.hpp>
 #include <jank/runtime/core/munge.hpp>
+#include <jank/runtime/obj/nil.hpp>
 #include <jank/runtime/obj/native_function_wrapper.hpp>
 #include <jank/runtime/obj/persistent_hash_map.hpp>
 #include <jank/runtime/obj/keyword.hpp>
+#include <jank/runtime/rtti.hpp>
 #include <jank/evaluate.hpp>
 #include <jank/codegen/llvm_processor.hpp>