--- a/compiler+runtime/src/cpp/jank/runtime/core/to_string.cpp
+++ b/compiler+runtime/src/cpp/jank/runtime/core/to_string.cpp
@@ -1,4 +1,3 @@
-#include <jank/native_persistent_string/fmt.hpp>
 #include <jank/runtime/core/to_string.hpp>
 #include <jank/runtime/visit.hpp>