## Changes
* Remove fmt entirely (this supersedes #298)
* Move equality fns into a new `core/equal` ns
* Use more extern templates
* Minimize includes
* folly submodule is updated to include a patch which replaces fmt usage with `std::format`
* immer submodule is updated with changes reflected in this PR I also made upstream: https://github.com/arximboldi/immer/pull/297

Lots more work to be done, but removing fmt helps with portability, too. Folly and BppTree are next on the chopping block, which should both have a huge impact on build times.

## Before
```
Compilation (112 times):
  Parsing (frontend):          890.0 s
  Codegen & opts (backend):     68.9 s
```

## After
```
Compilation (116 times):
  Parsing (frontend):          800.5 s
  Codegen & opts (backend):     59.3 s
```