@frenchy64 2025-03-13
Seeing this on my Intel mac:

```
[105/230] Building CXX object CMakeFiles/jank_lib.dir/src/cpp/jank/runtime/context.cpp.o
FAILED: CMakeFiles/jank_lib.dir/src/cpp/jank/runtime/context.cpp.o
/usr/local/opt/llvm/bin/clang++ -D__STDC_CONSTANT_MACROS -D__STDC_CONSTANT_MACROS -D__STDC_FORMAT_MACROS -D__STDC_LIMIT_MACROS -D__STDC_FORMAT_MACROS -D__STDC_LIMIT_MACROS -I/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/build -I/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime -I/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp -isystem /usr/local/opt/llvm/include -isystem /usr/local/include -isystem /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/third-party/nanobench/include -isystem /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/third-party/folly -isystem /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/third-party/bpptree/include -isystem /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/third-party/immer -isystem /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/third-party/magic_enum/include/magic_enum -isystem /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/third-party/cli11/include -isystem /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/third-party/ftxui/include -isystem /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/third-party/libzippp/src -isystem /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/third-party/boost-preprocessor/include -isystem /usr/local/Cellar/openssl@3/3.4.1/include -isystem /usr/local/Cellar/llvm/19.1.7_1/include/c++/v1 -isystem /usr/local/Cellar/llvm/19.1.7_1/lib/clang/19/include -isystem /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX15.2.sdk/usr/include -fPIC -fvisibility-inlines-hidden -Werror=date-time -Werror=unguarded-availability-new -Wall -Wextra -Wno-unused-parameter -Wwrite-strings -Wcast-qual -Wmissing-field-initializers -Wimplicit-fallthrough -Wcovered-switch-default -Wno-noexcept-type -Wnon-virtual-dtor -Wdelete-non-virtual-dtor -Wsuggest-override -Wstring-conversion -Wmisleading-indentation -Wctad-maybe-unsupported -fdiagnostics-color -g -std=gnu++20 -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX15.2.sdk -fcolor-diagnostics -std=gnu++20 -DIMMER_HAS_LIBGC=1 -DIMMER_TAGGED_NODE=0 -DHAVE_CXX14=1 -Werror -DJANK_DEBUG -DJANK_TEST -Wall -Wextra -Wpedantic -Wfloat-equal -Wuninitialized -Wswitch-enum -Wnon-virtual-dtor -Wold-style-cast -Wno-gnu-case-range -Wno-gnu-conditional-omitted-operand -Wno-implicit-fallthrough -Wno-covered-switch-default -Wno-invalid-offsetof -Wno-deprecated-declarations -Wno-c++23-extensions -fno-common -fno-rtti -fexceptions -O0 -DJANK_VERSION=\"jank-0.1-296d2559b38d751ddb73e7065f2eea5d5afb95a9\" "-DJANK_JIT_FLAGS=\"-std=gnu++20 -DIMMER_HAS_LIBGC=1 -DIMMER_TAGGED_NODE=0 -DHAVE_CXX14=1 -Werror -DJANK_DEBUG -Werror -DJANK_TEST -w\"" -DJANK_CLANG_PREFIX=\"/usr/local/opt/llvm\" -MD -MT CMakeFiles/jank_lib.dir/src/cpp/jank/runtime/context.cpp.o -MF CMakeFiles/jank_lib.dir/src/cpp/jank/runtime/context.cpp.o.d -o CMakeFiles/jank_lib.dir/src/cpp/jank/runtime/context.cpp.o -c /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/src/cpp/jank/runtime/context.cpp
In file included from /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/src/cpp/jank/runtime/context.cpp:25:
In file included from /Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/fmt/print.hpp:3:
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/fmt.hpp:18:9: error: call to object of type 'string_builder' is ambiguous
   18 |         sb(std::forward<T>(arg));
      |         ^~
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/fmt.hpp:31:40: note: in instantiation of function template specialization 'jank::util::format<unsigned long long>' requested here
   31 |     int const dummy[sizeof...(Args)]{ (format(sb, fmt, std::forward<Args>(args)), 0)... };
      |                                        ^
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/src/cpp/jank/runtime/context.cpp:372:18: note: in instantiation of function template specialization 'jank::util::format<jank::native_persistent_string, const char *, unsigned long long>' requested here
  372 |     return util::format("{}-{}-{}",
      |                  ^
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/string_builder.hpp:23:21: note: candidate function
   23 |     string_builder &operator()(native_bool d) &;
      |                     ^
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/string_builder.hpp:24:21: note: candidate function
   24 |     string_builder &operator()(native_integer d) &;
      |                     ^
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/string_builder.hpp:25:21: note: candidate function
   25 |     string_builder &operator()(native_real d) &;
      |                     ^
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/string_builder.hpp:26:21: note: candidate function
   26 |     string_builder &operator()(native_hash d) &;
      |                     ^
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/string_builder.hpp:28:21: note: candidate function
   28 |     string_builder &operator()(int d) &;
      |                     ^
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/string_builder.hpp:29:21: note: candidate function
   29 |     string_builder &operator()(long d) &;
      |                     ^
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/string_builder.hpp:30:21: note: candidate function
   30 |     string_builder &operator()(size_t d) &;
      |                     ^
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/string_builder.hpp:31:21: note: candidate function
   31 |     string_builder &operator()(char d) &;
      |                     ^
/Users/ambrose/Projects/jank-local-dev/origin-compile_times/compiler+runtime/include/cpp/jank/util/string_builder.hpp:32:21: note: candidate function
   32 |     string_builder &operator()(char32_t d) &;
      |                     ^
1 error generated.
[114/230] Building CXX object CMakeFiles/jank_lib.dir/src/cpp/jank/runtime/obj/persistent_sorted_set.cpp.o
ninja: build stopped: subcommand failed.
```